Synchronize exercise round grades in the gradebook correctly.
There were bugs that caused the exercise round grades in the
gradebook to fall out of sync. If the course structure was modified
in a way that changed the max points of a round, the round would
display incorrect points because it could still include points
from, e.g., deleted exercises. Furthermore, Moodle scales the
grades everytime the max points change, and the scaling usually
results in incorrect grades in this case.

The solution is to insert exercise round grades for each student
again to the gradebook when the max points change.
An alternative solution is to use gradebook formulas that would
make Moodle compute the round grades automatically based on the
defined dependencies. That could be considered in the future.