Reload randomized questionnaires after submitting in chapters
The exercise description should be reloaded after submitting
because it may change in personalized or randomized exercises.
This had been previously implemented in chapter.js for
asynchronously graded exercises, but not quizzes. Usually, quizzes
show the feedback in place of the exercise description (and thus
their exercise descriptions need not be reloaded), but that
can not be done for randomized quizzes since their nonces are not
reusable. Thus, randomized quizzes should not have the
data-aplus-quiz attribute, in which case chapter.js treats them
like asynchronous exercises. This commit reloads the exercise
description of randomized questionnaires (or any non-quiz exercise
that does not poll for the feedback) after submitting.

Randomized quizzes were not previously reloaded after submitting
because the reloading was only done in the callback function for
asynchronous waiting for the feedback.