Remove exercise grade items from the gradebook
From now on, the gradebook contains only the total grade for
each exercise round.

Moodle 3.8 changed the gradebook API so that it is not feasible
to have a variable number of grade items per activity instance.
Therefore, it was best to remove the exercise grade items
completely. Each exercise round may have a different number of
exercises. The first Astra release for Moodle 3.8 still had
the exercise grade items, but there was a hard limit for
the number of exercises in a round (30). Now there is no limit
again.

Add a CLI script for removing exercise grade items from
the gradebook. It should be run before upgrading to Moodle 3.9
and before upgrading the Astra plugin to this commit's version
since this commit already removes the gradeitemnumber field
from the exercise table.